// src/API/useCases/ADMIN/customer/entity.Customer.ts
export interface EntityCustomer {
    Delete?: boolean;
    id?: string;
    Name?: string;
    Mobile?: string;
    Email?: string;
    DOB?: string;
    Address?: string;
    ZipCode?: string;
    Points?: string;
    SMSAndEmailPromotions?: boolean;
    HouseAccEligible?: boolean;
    MaxAmountOfBalance?: string
    
}